type pinboard-api
